split optimisation into multiple modules and put in separate sub-dir
